requirements.txt
feedparser
aiohttp
aiofiles
pyyaml
beautifulsoup4